# internal/app/openapi.yaml
openapi: 3.1.0
info:
  title: WaCalls API
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
  description: >-
    HTTP API for WaCalls, a self-hosted WhatsApp calling bridge. All /api/*
    endpoints require the bearer token configured via WACALLS_API_TOKEN
    (when unset, authentication is disabled). The SSE stream also accepts the
    token as the access_token query parameter. This document is the canonical
    contract: a test gate fails the build when routes and spec drift apart.
servers:
  - url: /
security:
  - bearerAuth: []
  - queryToken: []
paths:
...
  /api/sessions/{sid}/onwhatsapp:
    post:
      operationId: isOnWhatsApp
      summary: Check which of the given phone numbers are registered on WhatsApp
      description: >-
        Added by the fork. Dialling a number without WhatsApp burns an attempt,
        raises the chip's failure rate and worsens how the account looks to Meta,
        so a list importer should check before queueing.


        Returns one result per number asked, in the order they were sent. The
        underlying USync query only answers about the users the server knows, so
        aligning results by position would silently put one contact's answer on
        another's row.


        Numbers may be sent with or without the leading `+`. At most 50 per
        request: the query goes to Meta's servers and volume draws attention, so
        an importer is expected to space its batches out and spread them across
        the tenant's chips.
      parameters:
        - $ref: "#/components/parameters/Sid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phones]
              properties:
                phones:
                  type: array
                  minItems: 1
                  maxItems: 50
                  items: { type: string }
              example:
                phones: ["5514981120008", "5514996035887"]
      responses:
        "200":
          description: One result per number asked, in the order they were sent
          content:
            application/json:
              schema:
                type: object
                required: [results]
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      required: [phone, exists]
                      properties:
                        phone:
                          type: string
                          description: The number exactly as it was sent
                        exists: { type: boolean }
                        jid:
                          type: string
                          description: Canonical JID, absent when the number is not on WhatsApp
              example:
                results:
                  - phone: "5514981120008"
                    exists: true
                    jid: "5514981120008@s.whatsapp.net"
                  - phone: "5514996035887"
                    exists: false
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: No such session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: The session has no WhatsApp account linked yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429": { $ref: "#/components/responses/RateLimited" }
        "502":
          description: WhatsApp did not answer the query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    queryToken:
      type: apiKey
      in: query
      name: access_token
  parameters:
    Sid:
      name: sid
      in: path
      required: true
      schema:
        type: string
      description: Session id
...
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Per-IP rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    BadRequest:
      description: Invalid request parameter or body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...